Revert "Create account view controller"
