--- .idea/dictionaries/school.xml
@@ -1,3 +1,0 @@
-<component name="ProjectDictionaryState">
-  <dictionary name="school" />
-</component>
