+++ src/main/java/com/example/werkplekkenfrontend/HelloController.java
@@ -1,0 +1,14 @@
+package com.example.werkplekkenfrontend;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Label;
+
+public class HelloController {
+    @FXML
+    private Label welcomeText;
+
+    @FXML
+    protected void onHelloButtonClick() {
+        welcomeText.setText("Welcome to JavaFX Application!");
+    }
+}
