--- src/main/java/com/example/werkplekkenfrontend/controllers/ViewController.java
@@ -1,8 +1,0 @@
-package com.example.werkplekkenfrontend.controllers;
-
-import com.example.werkplekkenfrontend.Main;
-import javafx.scene.Scene;
-
-public interface ViewController {
-    void updateView();
-}
