--- src/main/java/com/example/werkplekkenfrontend/controllers/canError.java
@@ -1,5 +1,0 @@
-package com.example.werkplekkenfrontend.controllers;
-
-public interface canError {
-    default void showMessageFailed(String message) {};
-}
