+++ src/main/java/module-info.java
@@ -6,6 +6,4 @@ module com.example.werkplekkenfrontend {
 
     opens com.example.werkplekkenfrontend to javafx.fxml;
     exports com.example.werkplekkenfrontend;
-    exports com.example.werkplekkenfrontend.controllers;
-    opens com.example.werkplekkenfrontend.controllers to javafx.fxml;
 }
