--- src/main/resources/com/example/werkplekkenfrontend/account-view.fxml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.layout.VBox?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.layout.HBox?>
-
-<?import javafx.scene.control.Label?>
-
-<VBox fx:id="main_container" minWidth="1080" minHeight="1080" alignment="TOP_CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml" fx:controller="com.example.werkplekkenfrontend.controllers.AccountViewController">
-    <padding>
-        <Insets top="100"/>
-    </padding>
-    <VBox maxWidth="800" prefHeight="400" spacing="50" alignment="CENTER_LEFT" style="-fx-border-color: black; -fx-padding: 50">
-        <HBox alignment="CENTER_LEFT" spacing="50">
-            <Label fx:id="firstName" maxHeight="20" prefWidth="200" style="-fx-border-color: black; -fx-padding: 50"/>
-            <Label fx:id="lastName" maxHeight="50" prefWidth="200" style="-fx-border-color: black; -fx-padding: 50"/>
-        </HBox>
-        <Label fx:id="email" prefWidth="400" prefHeight="50" style="-fx-border-color: black; -fx-padding: 50"/>
-    </VBox>
-    <Button text="Log Out" onAction="#onLogOutClick" prefHeight="50" maxWidth="800"/>
-</VBox>
-
