--- src/main/resources/com/example/werkplekkenfrontend/admin-edit-building-view.fxml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-        <?import javafx.scene.layout.VBox?>
-
-        <?import javafx.scene.control.Button?>
-        <?import javafx.scene.layout.HBox?>
-
-<?import javafx.scene.control.TextArea?>
-<VBox minWidth="1080" minHeight="1080" alignment="TOP_CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml" fx:controller="com.example.werkplekkenfrontend.controllers.AdminEditBuildingViewController">
-    <VBox alignment="TOP_CENTER" >
-        <HBox prefWidth="1080" prefHeight="380" alignment="CENTER" spacing="200" xmlns:fx="http://javafx.com/fxml">
-            <Button text="Cancel" onAction="#onCancelClick" cancelButton="true" prefHeight="50" prefWidth="200"/>
-            <Button text="Apply" onAction="#onApplyClick" defaultButton="true" prefHeight="80" prefWidth="80"/>
-        </HBox>
-        <VBox maxWidth="800" maxHeight="400">
-            <TextArea fx:id="name"/>
-            <HBox>
-                <TextArea fx:id="zipcode"/>
-                <TextArea fx:id="city"/>
-            </HBox>
-            <TextArea fx:id="adress"/>
-        </VBox>
-    </VBox>
-
-</VBox>
