--- src/main/resources/com/example/werkplekkenfrontend/admin-view.fxml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.control.Button?>
-
-<VBox fx:id="main_container" minWidth="1080" minHeight="1080" alignment="TOP_CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml" fx:controller="com.example.werkplekkenfrontend.controllers.AdminViewController">
-    <Button text="Buildings" onAction="#onBuildingsButtonClick" prefHeight="300" prefWidth="800"/>
-    <Button text="Roles" onAction="#onRolesButtonClick" prefHeight="300" prefWidth="800"/>
-</VBox>
